requirements.txt
git+https://github.com/Rapptz/discord.py/
mysql-connector-python==8.0.29